refactor: move model validators and enums
