+++ pokemon/models/__init__.py
@@ -1,11 +1,9 @@
 """Convenience re-exports for Pokémon models."""
 
+from .validators import validate_ivs, validate_evs
+from .enums import Gender, Nature
 from .core import (
     MAX_PP_MULTIPLIER,
-    validate_ivs,
-    validate_evs,
-    Gender,
-    Nature,
     SpeciesEntry,
     BasePokemon,
     Pokemon,
